refactor(persistence): extract sidecar load resolution
